+++ src/utils/index.ts
@@ -4,6 +4,7 @@ import { MFE, MFME } from './messages'
 import { PATA } from './constants'
 import { decryptMessage, encryptMessage, generatePassword } from './security'
 import { response } from './response'
+import { signAccessToken } from './jwt.helper'
 
 export {
   docs,
@@ -14,5 +15,6 @@ export {
   decryptMessage,
   encryptMessage,
   generatePassword,
-  response
+  response,
+  signAccessToken
 }